+++ RELEASENOTES.md
@@ -4,6 +4,8 @@
 
 * HLS:
   * Form an adaptive track group out of audio renditions with matching name.
+  * Support encrypted initialization segments
+    ([#5441](https://github.com/google/ExoPlayer/issues/5441)).
 * DASH:
   * Fix issue handling large `EventStream` presentation timestamps
     ([#5490](https://github.com/google/ExoPlayer/issues/5490)).